make Ports a final class
