+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - `Innmind\OperatingSystem\OperatingSystem` is now a final class
 - `Innmind\OperatingSystem\Sockets` is now a final class
 - `Innmind\OperatingSystem\Remote` is now a final class
+- `Innmind\OperatingSystem\Ports` is now a final class
 
 ### Fixed
