--- src/Ports/Unix.php
@@ -1,42 +1,0 @@
-<?php
-declare(strict_types = 1);
-
-namespace Innmind\OperatingSystem\Ports;
-
-use Innmind\OperatingSystem\{
-    Ports,
-    Config,
-};
-use Innmind\Url\Authority\Port;
-use Innmind\IO\Sockets\Internet\Transport;
-use Innmind\IP\IP;
-use Innmind\Immutable\Attempt;
-
-final class Unix implements Ports
-{
-    private Config $config;
-
-    private function __construct(Config $config)
-    {
-        $this->config = $config;
-    }
-
-    /**
-     * @internal
-     */
-    public static function of(Config $config): self
-    {
-        return new self($config);
-    }
-
-    #[\Override]
-    public function open(Transport $transport, IP $ip, Port $port): Attempt
-    {
-        return $this
-            ->config
-            ->io()
-            ->sockets()
-            ->servers()
-            ->internet($transport, $ip, $port);
-    }
-}
